+++ hadoop-ozone/recon/src/main/java/org/apache/hadoop/ozone/recon/api/handlers/RootEntityHandler.java
@@ -1,10 +1,31 @@
-package org.apache.hadoop.ozone.recon.api;
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.hadoop.ozone.recon.api.handlers;
 
 import org.apache.hadoop.hdds.scm.server.OzoneStorageContainerManager;
 import org.apache.hadoop.ozone.om.helpers.OmBucketInfo;
 import org.apache.hadoop.ozone.om.helpers.OmVolumeArgs;
 import org.apache.hadoop.ozone.recon.ReconConstants;
-import org.apache.hadoop.ozone.recon.api.types.*;
+import org.apache.hadoop.ozone.recon.api.types.NamespaceSummaryResponse;
+import org.apache.hadoop.ozone.recon.api.types.EntityType;
+import org.apache.hadoop.ozone.recon.api.types.DUResponse;
+import org.apache.hadoop.ozone.recon.api.types.QuotaUsageResponse;
+import org.apache.hadoop.ozone.recon.api.types.FileSizeDistributionResponse;
 import org.apache.hadoop.ozone.recon.recovery.ReconOMMetadataManager;
 import org.apache.hadoop.ozone.recon.spi.ReconNamespaceSummaryManager;
 
@@ -12,6 +33,9 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * Class for handling root entity type.
+ */
 public class RootEntityHandler extends EntityHandler {
   public RootEntityHandler(
       ReconNamespaceSummaryManager reconNamespaceSummaryManager,
@@ -23,8 +47,8 @@ public class RootEntityHandler extends EntityHandler {
   @Override
   public NamespaceSummaryResponse getSummaryResponse(String[] names) throws
       IOException {
-    NamespaceSummaryResponse namespaceSummaryResponse = new NamespaceSummaryResponse(
-        EntityType.ROOT);
+    NamespaceSummaryResponse namespaceSummaryResponse =
+            new NamespaceSummaryResponse(EntityType.ROOT);
     List<OmVolumeArgs> volumes = listVolumes();
     namespaceSummaryResponse.setNumVolume(volumes.size());
     List<OmBucketInfo> allBuckets = listBucketsUnderVolume(null);
@@ -87,7 +111,8 @@ public class RootEntityHandler extends EntityHandler {
   }
 
   @Override
-  public QuotaUsageResponse getQuotaResponse(String[] names) throws IOException {
+  public QuotaUsageResponse getQuotaResponse(String[] names)
+          throws IOException {
     QuotaUsageResponse quotaUsageResponse = new QuotaUsageResponse();
     List<OmVolumeArgs> volumes = listVolumes();
     List<OmBucketInfo> buckets = listBucketsUnderVolume(null);
@@ -115,7 +140,8 @@ public class RootEntityHandler extends EntityHandler {
   }
 
   @Override
-  public FileSizeDistributionResponse getDistResponse(String[] names) throws IOException {
+  public FileSizeDistributionResponse getDistResponse(String[] names)
+          throws IOException {
     FileSizeDistributionResponse distResponse =
         new FileSizeDistributionResponse();
     List<OmBucketInfo> allBuckets = listBucketsUnderVolume(null);
